src/cloud_functions/bronze/requirements.txt
requests==2.32.5
google-cloud-storage==3.7.0